regex only removes carret and dolar
