+++ tests/Mutator/Regex/PregMatchRemoveDollarTest.php
@@ -14,7 +14,7 @@ use Infection\Tests\Mutator\AbstractMutatorTestCase;
 /**
  * @internal
  */
-final class PregMatchSwapDollarTest extends AbstractMutatorTestCase
+final class PregMatchRemoveDollarTest extends AbstractMutatorTestCase
 {
     /**
      * @dataProvider provideMutationCases
@@ -26,20 +26,6 @@ final class PregMatchSwapDollarTest extends AbstractMutatorTestCase
 
     public function provideMutationCases(): \Generator
     {
-        yield 'It mutates correctly adding dollar when provided with a string' => [
-            <<<'PHP'
-<?php
-
-preg_match('#some-regexp#ig', 'irrelevant');
-PHP
-            ,
-            <<<'PHP'
-<?php
-
-preg_match('#some-regexp$#ig', 'irrelevant');
-PHP
-        ];
-
         yield 'It mutates correctly removing dollar when provided with a string' => [
             <<<'PHP'
 <?php
